+++ packages/node/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update `@subql/node-core` (#383)
+
+### Added
+- Support for ERC-4844 fields and added transaction type filter (#383)
 
 ## [6.0.5] - 2025-05-08
 ### Changed